+++ backend/packages/gitlab/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cat-factory/gitlab
 
+## 0.7.44
+
+### Patch Changes
+
+- Updated dependencies [51869b8]
+  - @cat-factory/kernel@0.116.0
+
 ## 0.7.43
 
 ### Patch Changes